requirements.txt
# HTTP Requests
requests>=2.31.0
fake-useragent>=1.4.0

# Data Processing
pandas>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
